# Removes unused FunctionInstance.LocalTypes