This reduces the memory pressure at instantiation slightly 
#889 

```
$ benchstat old.txt new.txt

name                                    old time/op    new time/op    delta
Initialization/interpreter-10             35.5µs ± 1%    34.5µs ± 1%  -2.73%  (p=0.000 n=10+8)
Initialization/interpreter-multiple-10    13.7µs ± 4%    13.7µs ± 8%    ~     (p=0.684 n=10+10)
Initialization/compiler-10                27.1µs ±49%    24.9µs ±49%    ~     (p=0.165 n=10+10)
Initialization/compiler-multiple-10       12.3µs ± 5%    12.1µs ± 4%    ~     (p=0.247 n=10+10)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-10              141kB ± 0%     139kB ± 0%  -1.27%  (p=0.000 n=10+10)
Initialization/interpreter-multiple-10     142kB ± 0%     140kB ± 0%  -1.26%  (p=0.000 n=8+8)
Initialization/compiler-10                 144kB ± 0%     142kB ± 0%  -1.25%  (p=0.000 n=9+9)
Initialization/compiler-multiple-10        144kB ± 0%     143kB ± 0%  -1.24%  (p=0.000 n=8+10)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-10               86.0 ± 0%      86.0 ± 0%    ~     (all equal)
Initialization/interpreter-multiple-10      92.0 ± 0%      92.0 ± 0%    ~     (all equal)
Initialization/compiler-10                  35.0 ± 0%      35.0 ± 0%    ~     (all equal)
Initialization/compiler-multiple-10         41.0 ± 0%      41.0 ± 0%    ~     (all equal)



```